Added initial stab for projecting CLR to JavaScript.
 - Still needs to support removeEventListener and oneventname

Added support for events.  However, to make sure that events work across
threads, we capture the SynchronizationContext.  As a result, the
developer must call `engine.AcquireContext()`.  Whenever an event is
registered from a host-provided object in JavaScript, the current
synchronization context is captured, and will be used to invoke it.  If no
synchronization context is current when the delegate is called, we will
attempt to acquire the context on the current thread.