Host/Object bridge: Properties implemented
This change alters the implementation for object bridges so that the
responsibility isn't 100% on JavaScriptConverter, but rather on a class
called BridgeManager and ClassBridge.  ClassBridge is main unit of work
which enumerates publics and creates the projected JavaScript types.
JavaScriptConverter then only has to do a little work to find the right
bridge and hook it up to an external object.

This first version implements properties support.  Properties should now
be faster with the new mechanism, because they compile dynamic methods
instead of using the reflection get/set mechanisms.